Refactor logging / exception handling code such that `import acq4` no longer affects exceptions
